added oop inventory task
This task uses object-oriented programming and functions to read an external text file, inventory.txt, to create shoe objects which can be viewed and updated. Additional shoe objects can also be created by the user and all changes are then rewritten back to the inventory.txt file. The menu format allows for an interactive and well-formatted experience and structured, readable code that is sectioned in 3 parts: the class & methods, functions and the menu.